- Создан отдельный синглтон класс DBStore, где создается persistent container - Заменены оставшиеся force unwrap
